split up big array
